Removed TextureManager, all TPK methods were moved to be its internal methods instead.

+++ Nikki/Reflection/Enum/eTPKVersion.cs
@@ -7,7 +7,7 @@ namespace Nikki.Reflection.Enum
 	/// <summary>
 	/// Enum of <see cref="TPKBlock"/> version types.
 	/// </summary>
-	public enum eTPKVersion : byte
+	public enum eTPKVersion : int
 	{
 		/// <summary>
 		///
